+++ src/Models/JobExecutionLogProxy.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Konekt\History\Models;
+
+use Konekt\Concord\Proxies\ModelProxy;
+
+class JobExecutionLogProxy extends ModelProxy
+{
+}